Added display modes command line support
