+++ README.md
@@ -259,6 +259,9 @@ This contains a list of alternate core setups. These take existing cores and mak
     -i, --info                Shows the details for the specified 'name'
     -l, --list                Lists out all of the values for 'name' and their details
 
+  display-modes            Enable all Display Modes
+    -p, --path                Absolute path to install location
+
   update-self              Check for updates to pupdate
 
   help                     Display more information on a specific command.
