+++ src/options/DisplayModesOptions.cs
@@ -1,0 +1,6 @@
+using CommandLine;
+
+namespace Pannella.Options;
+
+[Verb("display-modes",  HelpText = "Enable all Display Modes")]
+public class DisplayModesOptions : BaseOptions { }
